moved the hardcoded admins into their own files. (still a bit of related logic on the rust file)
